Separated Gulp tasks into separate files
