+++ tasks/gulp/csslint.js
@@ -1,0 +1,14 @@
+var gulp = require('gulp'),
+  plugins = require('gulp-load-plugins')(),
+  defaultAssets = require('../../config/assets/default');
+
+gulp.task('csslint', function (done) {
+  return gulp.src(defaultAssets.client.css)
+    .pipe(plugins.csslint('.csslintrc'))
+    .pipe(plugins.csslint.reporter())
+    .pipe(plugins.csslint.reporter(function (file) {
+      if (!file.csslint.errorCount) {
+        done();
+      }
+    }));
+});
