+++ tasks/gulp/environment.js
@@ -1,0 +1,16 @@
+var gulp = require('gulp');
+
+// Set NODE_ENV to 'test'
+gulp.task('env:test', function () {
+  process.env.NODE_ENV = 'test';
+});
+
+// Set NODE_ENV to 'development'
+gulp.task('env:dev', function () {
+  process.env.NODE_ENV = 'development';
+});
+
+// Set NODE_ENV to 'production'
+gulp.task('env:prod', function () {
+  process.env.NODE_ENV = 'production';
+});
