+++ tasks/gulp/less.js
@@ -1,0 +1,14 @@
+var gulp = require('gulp'),
+  plugins = require('gulp-load-plugins')(),
+  defaultAssets = require('../../config/assets/default');
+
+
+gulp.task('less', function () {
+  return gulp.src(defaultAssets.client.less)
+    .pipe(plugins.less())
+    .pipe(plugins.autoprefixer())
+    .pipe(plugins.rename(function (file) {
+      file.dirname = file.dirname.replace(path.sep + 'less', path.sep + 'css');
+    }))
+    .pipe(gulp.dest('./modules/'));
+});
