+++ tasks/gulp/localConfig.js
@@ -1,0 +1,18 @@
+var gulp = require('gulp'),
+  plugins = require('gulp-load-plugins')(),
+  fs = require('fs');
+
+// Copy local development environment config
+gulp.task('copy:localConfig', function () {
+  var src = [];
+  var renameTo = 'local-development.js';
+
+  // only add the copy source if our destination file doesn't already exist
+  if (!fs.existsSync('config/env/' + renameTo)) {
+    src.push('config/env/local.example.js');
+  }
+
+  return gulp.src(src)
+    .pipe(plugins.rename(renameTo))
+    .pipe(gulp.dest('config/env'));
+});
