+++ tasks/gulp/protractor.js
@@ -1,0 +1,29 @@
+var gulp = require('gulp'),
+  protractor = require('gulp-protractor').protractor,
+  webdriver_update = require('gulp-protractor').webdriver_update,
+  webdriver_standalone = require('gulp-protractor').webdriver_standalone;
+
+// Downloads the selenium webdriver
+gulp.task('webdriver_update', webdriver_update);
+
+// Start the standalone selenium server
+// NOTE: This is not needed if you reference the
+// seleniumServerJar in your protractor.conf.js
+gulp.task('webdriver_standalone', webdriver_standalone);
+
+// Protractor test runner task
+gulp.task('protractor', ['webdriver_update'], function () {
+  gulp.src([])
+    .pipe(protractor({
+      configFile: 'protractor.conf.js'
+    }))
+    .on('end', function() {
+      console.log('E2E Testing complete');
+      // exit with success.
+      process.exit(0);
+    })
+    .on('error', function(err) {
+      console.log('E2E Tests failed');
+      process.exit(1);
+    });
+});
