+++ tasks/gulp/templatecache.js
@@ -1,0 +1,27 @@
+var gulp = require('gulp'),
+  plugins = require('gulp-load-plugins')({
+    rename: {
+      'gulp-angular-templatecache': 'templateCache'
+    }
+  }),
+  path = require('path'),
+  endOfLine = require('os').EOL,
+  defaultAssets = require('../../config/assets/default');
+
+// Angular template cache task
+gulp.task('templatecache', function () {
+  var re = new RegExp('\\' + path.sep + 'client\\' + path.sep, 'g');
+
+  return gulp.src(defaultAssets.client.views)
+    .pipe(plugins.templateCache('templates.js', {
+      root: 'modules/',
+      module: 'core',
+      templateHeader: '(function () {' + endOfLine + '	\'use strict\';' + endOfLine + endOfLine + '	angular' + endOfLine + '		.module(\'<%= module %>\'<%= standalone %>)' + endOfLine + '		.run(templates);' + endOfLine + endOfLine + '	templates.$inject = [\'$templateCache\'];' + endOfLine + endOfLine + '	function templates($templateCache) {' + endOfLine,
+      templateBody: '		$templateCache.put(\'<%= url %>\', \'<%= contents %>\');',
+      templateFooter: '	}' + endOfLine + '})();' + endOfLine,
+      transformUrl: function (url) {
+        return url.replace(re, path.sep);
+      }
+    }))
+    .pipe(gulp.dest('build'));
+});
